+++ src/com/pinktwins/elephant/util/Images.java
@@ -1,6 +1,8 @@
 package com.pinktwins.elephant.util;
 
+import java.awt.Graphics2D;
 import java.awt.Image;
+import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
@@ -42,5 +44,24 @@ public class Images {
 
 		String s = FilenameUtils.getExtension(f.getName()).toLowerCase();
 		return "png tif jpg jpeg bmp gif".indexOf(s) >= 0;
+	}
+	
+	// https://stackoverflow.com/questions/13605248/java-converting-image-to-bufferedimage
+	public static BufferedImage toBufferedImage(Image img)
+	{
+	    if (img instanceof BufferedImage) {
+	        return (BufferedImage) img;
+	    }
+
+	    // Create a buffered image with transparency
+	    BufferedImage bimage = new BufferedImage(img.getWidth(null), img.getHeight(null), BufferedImage.TYPE_INT_ARGB);
+
+	    // Draw the image on to the buffered image
+	    Graphics2D bGr = bimage.createGraphics();
+	    bGr.drawImage(img, 0, 0, null);
+	    bGr.dispose();
+
+	    // Return the buffered image
+	    return bimage;
 	}
 }